==> dataset_creation/3_prepare_dataset_for_prompts.ipynb <==
import os
import json
from tqdm.notebook import tqdm
from ast import literal_eval
import pandas as pd
from glob import glob

PATH_YOUTUBE_DATA = "../dataset/Dataset40k/youtube_data"
PATH_SPOTIFY_YOUTUBE_CSV = "../dataset/Dataset40k/spotify_with_youtube.csv"
PATH_SAVE_FINAL_CSV = "../dataset/Dataset40k/spotify_with_youtube_clean.csv"

assert os.path.exists(PATH_YOUTUBE_DATA) and os.path.isdir(PATH_YOUTUBE_DATA), \
    f"Received bad path `{PATH_YOUTUBE_DATA}`"
assert os.path.exists(PATH_SPOTIFY_YOUTUBE_CSV), f"Received bad path `{PATH_SPOTIFY_YOUTUBE_CSV}`"


# Load Spotify and YouTube csv

# Setup
df = pd.read_csv(PATH_SPOTIFY_YOUTUBE_CSV)
assert not df["track_id"].duplicated().any()
assert not df.isna().any().any()
to_list_columns = ["artists", "track_genre"]
for column in to_list_columns:
    df[column] = df[column].apply(literal_eval)

spotify_mapper = {
    'track_id': 'spotify_track_id',
    'track_name': 'spotify_track_name',
    'artists': 'spotify_artists',
    'album_name': 'spotify_album_name',
    'track_genre': 'spotify_genre',
    'popularity': 'spotify_popularity',
    'energy': 'spotify_energy',
    'valence': 'spotify_valence',
    'danceability': 'spotify_danceability',
    'tempo': 'spotify_tempo',
    'duration_ms': 'spotify_duration_ms',
    'key': 'spotify_key',
    'mode': 'spotify_mode',
    'loudness': 'spotify_loudness',
    'time_signature': 'spotify_time_signature',
    'acousticness': 'spotify_acousticness',
    'instrumentalness': 'spotify_instrumentalness',
    'speechiness': 'spotify_speechiness',
    'liveness': 'spotify_liveness',
    'explicit': 'spotify_explicit_content',
    'original_index': 'spotify_original_index',
}

df = df.rename(columns=spotify_mapper)
df.head()

# Remove rows without YouTube metadata

audio_paths = glob(f"{PATH_YOUTUBE_DATA}/*.mp3")
audio_paths_base_names = [os.path.basename(p)[:-4] for p in audio_paths]
has_youtube_data = df["spotify_track_id"].isin(audio_paths_base_names)
print(len(df))
df = df[has_youtube_data]
print(len(df))

# Append YouTube metadata
NOTE: Will take a couple of minutes

paths = glob(f"{PATH_YOUTUBE_DATA}/*")
for i, row in tqdm(df.iterrows(), total=len(df)):
    track_id = row["spotify_track_id"]
    track_paths = [p.replace("\\", "/").replace("//", "/") for p in paths if (track_id in p)]
    if not all(os.path.exists(p) for p in track_paths) or (len(track_paths) != 3):
        print(i, end=",")
        df = df.drop(index=i)
        continue

    # Json
    json_path = track_paths[0]
    assert json_path.endswith(".json")
    assert os.path.exists(json_path), f"Received bad path `{json_path}`"
    with open(json_path, 'r') as f:
        json_data = json.load(f)
    df.loc[i, "youtube_title_download"] = str(json_data["title"])
    df.loc[i, "youtube_duration_ms"] = int(json_data["length_ms"])
    df.loc[i, "youtube_views"] = int(json_data["views"])
    df.loc[i, "youtube_english_captions"] = str(json_data["captions_english"])

    # Audio
    audio_path = track_paths[1] 
    assert audio_path.endswith(".mp3")
    assert os.path.exists(audio_path), f"Received bad path `{audio_path}`"
    df.loc[i, "path_audio_full"] = audio_path

    # Thumbnail
    thumbnail_path = track_paths[2]
    assert thumbnail_path.endswith(".jpg")
    assert os.path.exists(thumbnail_path), f"Received bad path `{thumbnail_path}`"

# Wrapping up and save new csv

reorder = [
    'spotify_track_id', 'spotify_track_name', 'spotify_artists',
    'spotify_album_name', 'spotify_genre', 'spotify_popularity',
    'spotify_energy', 'spotify_valence', 'spotify_danceability',
    'spotify_tempo', 'spotify_duration_ms', 'spotify_key', 'spotify_mode',
    'spotify_loudness', 'spotify_time_signature', 'spotify_acousticness',
    'spotify_instrumentalness', 'spotify_speechiness', 'spotify_liveness',
    'spotify_explicit_content', 'spotify_original_index', 'youtube_url',
    'youtube_title', 'youtube_views', 'youtube_duration_ms',
    'youtube_video_skipped', 'youtube_search_query',
    'youtube_title_download', 'youtube_english_captions', "path_audio_full"
]
assert set(reorder) - set(df.columns.tolist()) == set(df.columns.tolist()) - set(reorder) == set()
df = df[reorder]
df = df.reset_index(drop=True)
df.to_csv(PATH_SAVE_FINAL_CSV, index=False)
df.head()
